Bumping up Elastic Search and Spring dependency versions (#991)
* Bumping up Elastic Search and Spring dependency versions

* Specifying the 'public' postgres schema

* Bumping up flyway version

* Declaring flyway schema

* Bumping chart version

* Trying to bypass the flyway check for preview

* Bump up chart version

* Disabling Spring native Kubernetes health endpoints

* Tidying up PR

+++ src/test/java/uk/gov/hmcts/reform/divorce/orchestration/event/listener/UpdateDNPronouncedCaseEventListenerTest.java
@@ -74,7 +74,7 @@ public class UpdateDNPronouncedCaseEventListenerTest {
     }
 
     private int casesProcessedCountInContext() {
-        return (int) context.getTransientObject(CASES_ELIGIBLE_FOR_DA_PROCESSED_COUNT);
+        return context.getTransientObject(CASES_ELIGIBLE_FOR_DA_PROCESSED_COUNT);
     }
 
     private void resetCassesProcessedCountInContext() {
